# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API de Tareas",
        "version": "1.0.0"
    },
    "paths": {
        "/tareas": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Obtener todas las tareas",
                "operationId": "52009680eb26b02127be9ebfb44a061c",
                "responses": {
                    "200": {
                        "description": "Lista de tareas",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Crear una nueva tarea",
                "operationId": "e0342625de01f5962370bf24e4e63256",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "nombre": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Tarea creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/tareas/{tarea}": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Obtener los detalles de una tarea",
                "operationId": "78a8907734f0a19a9a3e507d7784819c",
                "parameters": [
                    {
                        "name": "tarea",
                        "in": "path",
                        "description": "ID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la tarea",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Actualizar una tarea existente",
                "operationId": "73477406fae912c785a67b6d1653e935",
                "parameters": [
                    {
                        "name": "tarea",
                        "in": "path",
                        "description": "ID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "nombre": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Tarea actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Eliminar una tarea",
                "operationId": "4b1e84aa68b57c6cd84211a1434e3d31",
                "parameters": [
                    {
                        "name": "tarea",
                        "in": "path",
                        "description": "ID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Tarea eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}